fix: order filters
